+++ Classes/Taglib/Tags.h
@@ -37,8 +37,11 @@ struct FileData;
 	
 	//	Sort by
 	
+	NSString *albumSort;
 	NSString *artistSort;
-	
+	NSString *titleSort;
+	NSString *composerSort;
+	NSString *albumArtistSort;
 }
 
 /// @name Initializing an Tags Object
@@ -135,6 +138,11 @@ struct FileData;
 
 /// @name Finding Sort by  Metadata 
 
+@property (assign) NSString *albumSort;
 @property (assign) NSString *artistSort;
+@property (assign) NSString *titleSort;
+@property (assign) NSString *composerSort;
+@property (assign) NSString *albumArtistSort;
+
 
 @end